back/requirements.txt
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
psycopg2==2.8.5
